tests fixed
