+++ packages/core/core/__tests__/utils/nuxt/proxyUtils.spec.ts
@@ -1,0 +1,82 @@
+import * as utils from '../../../src/utils/nuxt/_proxyUtils';
+
+describe('[CORE - utils] _proxyUtils', () => {
+  it('returns proxy for defined api', () => {
+    const givenApi = {
+      getProduct: jest.fn()
+    };
+
+    const client = {
+      post: jest.fn(() => ({ then: jest.fn() }))
+    };
+
+    const proxiedApi = utils.createProxiedApi({ givenApi, client, tag: 'ct' });
+
+    proxiedApi.getProduct({ product: 1 });
+    proxiedApi.getCategory({ category: 1 });
+
+    expect(givenApi.getProduct).toBeCalled();
+    expect(client.post).toBeCalledWith('/ct/getCategory', [{ category: 1 }]);
+  });
+
+  it('reads cookies from incoming request', () => {
+    expect(utils.getCookies(null)).toEqual('');
+    expect(utils.getCookies({} as any)).toEqual('');
+    expect(utils.getCookies({ req: { headers: {} } } as any)).toEqual('');
+    expect(utils.getCookies({ req: { headers: { cookie: { someCookie: 1 } } } } as any)).toEqual({ someCookie: 1 });
+  });
+
+  it('it combines config with the current one', () => {
+    jest.spyOn(utils, 'getCookies').mockReturnValue('');
+
+    expect(utils.getIntegrationConfig(
+      {
+        $config: {
+          middlewareUrl: 'some-url'
+        }
+      } as any,
+      { someGivenOption: 1 }
+    )).toEqual({
+      axios: {
+        baseURL: 'some-url',
+        headers: {}
+      },
+      someGivenOption: 1
+    });
+  });
+
+  it('it combines config with the current one and adds a cookie', () => {
+    jest.spyOn(utils, 'getCookies').mockReturnValue('xxx');
+
+    expect(utils.getIntegrationConfig(
+      {
+        $config: {
+          middlewareUrl: 'some-url'
+        }
+      } as any,
+      {}
+    )).toEqual({
+      axios: {
+        baseURL: 'some-url',
+        headers: {
+          cookie: 'xxx'
+        }
+      }
+    });
+  });
+
+  it('it throws error when no middlewareUrl is passed in config', () => {
+    jest.spyOn(utils, 'getCookies').mockReturnValue('');
+
+    expect(() => {
+      utils.getIntegrationConfig(
+        {
+          $config: {
+            middlewareUrl: ''
+          }
+        } as any,
+        {}
+      );
+    }).toThrowError();
+  });
+});
